+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * 【http   】     增加HttpResource（issue#1943@Github）
 * 【http   】     增加BytesBody、FormUrlEncodedBody
 * 【cron   】     TaskTable.remove增加返回值（issue#I4HX3B@Gitee）
+* 【core   】     Tree增加filter、filterNew、cloneTree、hasChild方法（issue#I4HFC6@Gitee）
 ### 🐞Bug修复
 * 【core   】     修复FileResource构造fileName参数无效问题（issue#1942@Github）